refactor(blackjack): GameResult를 도메인 성격으로 보게되어 도메인 패키지로 이동
